完成更改防卫project模块的sql注入查

+++ creaction/service/project/modifyProject.py
@@ -80,15 +80,15 @@ def __updateProjectStorage(userUUID, projectUUID, dataJsonDict, mediasDict, memb
         Loger.error(MESSAGE[CODE_ERROR_IMAGE_NUMBER_TOO_MANY], __file__)
         return RESPONSE_JSON(CODE_ERROR_IMAGE_NUMBER_TOO_MANY)
     
-    oldMediasQuerySQL = """ SELECT media_name FROM t_project_media WHERE project_uuid='%s';""" % projectUUID
-    sqlList = __packageSQL(userUUID, projectUUID, dataJsonDict, mediasDict, memberList)
+    oldMediasQuerySQL = """ SELECT media_name FROM t_project_media WHERE project_uuid=%s;"""
+    sqlList, argsList = __packageSQL(userUUID, projectUUID, dataJsonDict, mediasDict, memberList)
 
     fileList = []
     dbManager = DB.DBManager.shareInstanced()
     try:
-        oldMediasList = dbManager.executeSingleQuery(oldMediasQuerySQL, False)
+        oldMediasList = dbManager.executeSingleQueryWithArgs(oldMediasQuerySQL, [projectUUID], False)
         fileList = [data[0] for data in oldMediasList]
-        dbManager.executeTransactionMutltiDml(sqlList)
+        dbManager.executeTransactionMutltiDmlWithArgsList(sqlList, argsList)
     except Exception as e:
         Loger.error(e, __file__)
         __removeFileOnError(path, mediasDict.values())
@@ -102,88 +102,124 @@ def __updateProjectStorage(userUUID, projectUUID, dataJsonDict, mediasDict, memb
 def __removeFileOnError(path, fileList):
     for fileName in fileList:
         fullPath = path + fileName
-        os.remove(fullPath)
+        if os.path.exists(fullPath):
+            os.remove(fullPath)
 
 
 def __packageSQL(userUUID, projectUUID, dataJsonDict, mediasDict, memberList):
     sqlList = []
+    argsList = []
     time = generateCurrentTime()
 
     # 项目多媒体内容(先删除后插入新的)
-    deleteMdiasSQL = """DELETE FROM t_project_media WHERE project_uuid='%s';""" % projectUUID
+    deleteMdiasSQL = """DELETE FROM t_project_media WHERE project_uuid=%s;"""
     sqlList.append(deleteMdiasSQL)
+    argsList.append([projectUUID])
+
     medias_count = len(mediasDict)
     if medias_count > 0:
+        insertMediasArgs = []
         insertMediasSQL = """INSERT INTO t_project_media (project_uuid, sorting, type, media_name, time) VALUES """
         values = ""
         typeInt = Config.TYPE_FOR_PROJECT_MEDIAS_PICTURE
         for key, value in mediasDict.items():
-            values += """('%s', %s, %d, '%s', '%s'),""" % (projectUUID, key, typeInt, value, time)
+            values += """(%s, %s, %s, %s, %s),"""
+            insertMediasArgs.append(projectUUID)
+            insertMediasArgs.append(key)
+            insertMediasArgs.append(str(typeInt))
+            insertMediasArgs.append(value)
+            insertMediasArgs.append(time)
         insertMediasSQL += values[:-1] + ";"
         sqlList.append(insertMediasSQL)
+        argsList.append(insertMediasArgs)
     
     # 项目计划列表(先删除后插入新的)
-    deletePlanSQL = """DELETE FROM t_project_plan WHERE project_uuid='%s';""" % projectUUID
+    deletePlanSQL = """DELETE FROM t_project_plan WHERE project_uuid=%s;"""
     sqlList.append(deletePlanSQL)
+    argsList.append([projectUUID])
+
     have_plan = 0
     planList = dataJsonDict["planList"]
     planListLen = len(planList)
     if planListLen > 0 :
         have_plan = 1
+        insertPlanArgs = []
         insertPlanSQL = """INSERT INTO t_project_plan (project_uuid, sorting, content, start_time, finish_time, medias_count) VALUES """
         values = ""
         for x in range(planListLen):
             data = planList[x]
-            content = data["content"]
-            startTime = data["startTime"]
-            finishTime = data["finishTime"]
-            values += """('%s', %d, '%s', '%s', '%s', 0),""" % (projectUUID, x, content, startTime, finishTime)
+            values += """(%s, %s, %s, %s, %s, 0),"""
+            insertPlanArgs.append(projectUUID)
+            insertPlanArgs.append(str(x))
+            insertPlanArgs.append(data["content"])
+            insertPlanArgs.append(data["startTime"])
+            insertPlanArgs.append(data["finishTime"])
         insertPlanSQL += values[:-1] + ";"
         sqlList.append(insertPlanSQL)
+        argsList.append(insertPlanArgs)
 
     # 项目标签列表(先删除后插入新的)
-    deleteTagsSQL = """DELETE FROM t_tag_project WHERE project_uuid='%s';""" % projectUUID
+    deleteTagsSQL = """DELETE FROM t_tag_project WHERE project_uuid=%s;"""
     sqlList.append(deleteTagsSQL)
+    argsList.append([projectUUID])
+
     tagList = dataJsonDict["tagList"]
     tagListLen = len(tagList)
     if tagListLen > 0 :
+        insertTagArgs = []
         insertTagSQL = """INSERT INTO t_tag_project (project_uuid, sorting, type, content) VALUES """
         values = ""
-        typeString = Config.TYPE_FOR_PROJECT_MEDIAS_PICTURE
         for i in range(tagListLen):
-            values += """('%s', %d, %d, '%s'),""" % (projectUUID, i, typeString, tagList[i])
+            values += """(%s, %s, %s, %s),"""
+            insertTagArgs.append(projectUUID)
+            insertTagArgs.append(str(i))
+            insertTagArgs.append(Config.TYPE_FOR_PROJECT_MEDIAS_PICTURE)
+            insertTagArgs.append(tagList[i])
         insertTagSQL += values[:-1] + ";"
         sqlList.append(insertTagSQL)
+        argsList.append(insertTagArgs)
 
     # 更新主项目
-    updateProjectSQL = """UPDATE t_project SET detail='{detail}', result='{result}', have_plan={have_plan}, medias_count={medias_count} WHERE uuid='{uuid}'; """.format(uuid=projectUUID, 
-        detail=dataJsonDict["detail"], result=dataJsonDict["result"], 
-        author_uuid=dataJsonDict["user_uuid"], have_plan=have_plan, 
-        medias_count=medias_count)
+    updateProjectArgs = []
+    updateProjectSQL = """UPDATE t_project SET detail=%s, result=%s, have_plan=%s, medias_count=%s WHERE uuid=%s; """
+    updateProjectArgs.append(dataJsonDict["detail"])
+    updateProjectArgs.append(dataJsonDict["result"])
+    updateProjectArgs.append(str(have_plan))
+    updateProjectArgs.append(str(medias_count))
+    updateProjectArgs.append(projectUUID)
     sqlList.append(updateProjectSQL)
+    argsList.append(updateProjectArgs)
 
     # 通知成员项目更新了
+    insertProjectMessageArgs = []
     insertProjectMessageSQL = """INSERT INTO t_message_project (user_uuid, type, content_uuid,
          owner_user_uuid, status, content, action, time) VALUES """
     values = ""
     typeString = Config.TYPE_FOR_MESSAGE_IN_PROJECT_UPDATE_PROJECT
     for member in memberList:
-        values += """('%s', %d, '%s', '%s', 0, '%s', 0, '%s'),""" % (member, typeString, projectUUID, userUUID, '更新了项目', time)
+        values += """(%s, %s, %s, %s, 0, %s, 0, %s),"""
+        insertProjectMessageArgs.append(member)
+        insertProjectMessageArgs.append(str(Config.TYPE_FOR_MESSAGE_IN_PROJECT_UPDATE_PROJECT))
+        insertProjectMessageArgs.append(projectUUID)
+        insertProjectMessageArgs.append(userUUID)
+        insertProjectMessageArgs.append('更新了项目')
+        insertProjectMessageArgs.append(time)
     insertProjectMessageSQL += values[:-1] + ";"
     sqlList.append(insertProjectMessageSQL)
+    argsList.append(insertProjectMessageArgs)
 
-    return sqlList
+    return sqlList, argsList
     
 
 def __queryProjectMembers(projectUUID):
     memberList = None
-    queryProjectMemberSQL = """SELECT user_uuid FROM t_project_user WHERE project_uuid='%s' AND type=%s""" % (projectUUID, Config.TYPE_FOR_PROJECT_MEMBER)
+    queryProjectMemberSQL = """SELECT user_uuid FROM t_project_user WHERE project_uuid=%s AND type=%s"""
+    queryProjectMemberArgs = [projectUUID, Config.TYPE_FOR_PROJECT_MEMBER]
 
     dbManager = DB.DBManager.shareInstanced()
     try:
-        tempList = dbManager.executeSingleQuery(queryProjectMemberSQL, False)
+        tempList = dbManager.executeSingleQueryWithArgs(queryProjectMemberSQL, queryProjectMemberArgs, False)
         memberList = [data[0] for data in tempList]
-
     except Exception as e:
         Loger.error(e, __file__)
         raise MemberQueryException()
